refactor(lexer): move to its own package
